Add files via upload
The complete elliptic integral of the first kind (CEI-1) plays in a significant role in mathematics, physics and engineering. There is no simple formulae for its computation, thus numerical algorithms and solution are essential in practical problems. However, we find that the numerical solutions obtained via both MATLAB and Mathematica are not acceptable and should be treated seriously. For the purpose of obtaining correct and replacable numerical algorithms for the CEI-1,  the infinite series method, arithmetic-geometric mean (AGM) method, Gauss-Chebyshev method and Gauss-Legendre methods are discussed in details with a top-down strategy. The four key algorithms for computing CEI-1 are designed, verified, validated and tested, which can be utilized in R & D and be reused properly. In the sense of STEM education, system engineering and computational thinking, the Verification-Validation-Testing (VVT) stage is crucial for applications and teaching college students in order to avoid unnecessary losses.